backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.2.5
pillow==11.0.0
numpy==2.1.2
face-recognition==1.3.0
python-multipart